Adds the state flow framework

+++ src/system/tests/system_state_logic_tests.cpp
@@ -33,17 +33,15 @@ StateId power_states[kNumPowerStates] = {kPowerEverythingOff, kPowerLimited,
 
 void TestPowerStateLogic() {
     BatteryChargeStateMachine battery_charge_state_machine(NULL);
-
     BatteryTempStateMachine battery_temp_state_machine(NULL);
-
     PowerStateMachine power_state_machine(NULL, &battery_charge_state_machine,
                                           &battery_temp_state_machine);
 
     // Check aaaaaaaall the permutations
-    for (int battery_charge_index = 0;
+    for (uint8_t battery_charge_index = 0;
          battery_charge_index < kNumBatteryChargeStates;
          battery_charge_index++) {
-        for (int battery_temp_index = 0;
+        for (uint8_t battery_temp_index = 0;
              battery_temp_index < kNumBatteryTempStates; battery_temp_index++) {
             battery_charge_state_machine.OverrideState(
                 battery_charge_states[battery_charge_index]);
@@ -83,15 +81,13 @@ void TestPowerStateLogic() {
 
 void TestTelecomsStateLogic() {
     PowerStateMachine power_state_machine(NULL, NULL, NULL);
-
     TelecomsTempStateMachine telecoms_temp_state_machine(NULL);
-
     TelecomsStateMachine telecoms_state_machine(NULL, &power_state_machine,
                                                 &telecoms_temp_state_machine);
 
     // Check aaaaaaaall the permutations
-    for (int power_index = 0; power_index < kNumPowerStates; power_index++) {
-        for (int telecoms_temp_index = 0;
+    for (uint8_t power_index = 0; power_index < kNumPowerStates; power_index++) {
+        for (uint8_t telecoms_temp_index = 0;
              telecoms_temp_index < kNumTelecomsTempStates;
              telecoms_temp_index++) {
             power_state_machine.OverrideState(power_states[power_index]);
@@ -126,20 +122,17 @@ void TestTelecomsStateLogic() {
 
 void TestAdcsStateLogic() {
     PowerStateMachine power_state_machine(NULL, NULL, NULL);
-
     TleStateMachine tle_state_machine(NULL);
-
     DetumbledStateMachine detumbled_state_machine(NULL);
-
     AdcsStateMachine adcs_state_machine(NULL, &power_state_machine,
                                         &tle_state_machine,
                                         &detumbled_state_machine);
 
     // Check aaaaaaaall the permutations
-    for (int power_index = 0; power_index < kNumPowerStates; power_index++) {
-        for (int detumbled_index = 0; detumbled_index < kNumDetumbledStates;
+    for (uint8_t power_index = 0; power_index < kNumPowerStates; power_index++) {
+        for (uint8_t detumbled_index = 0; detumbled_index < kNumDetumbledStates;
              detumbled_index++) {
-            for (int tle_index = 0; tle_index < kNumTleStates; tle_index++) {
+            for (uint8_t tle_index = 0; tle_index < kNumTleStates; tle_index++) {
                 power_state_machine.OverrideState(power_states[power_index]);
 
                 tle_state_machine.OverrideState(tle_states[tle_index]);
